Improve performance of header processing
According to the result of profiling, 'lhttpc_lib:header_value/2,3'
seems to be one of the hot spot. This commit tries to reduce the
overhead of calling this function.

To send a request, it can be skipped completely if a user passes new
flags 'is_host_defined' and 'is_content_length_defined'.
To receive a request, it's difficult to skip the call completely but the
performance of this function should be better by ~3 times.